broad testing improvements closes #2
With this change, the vast majority of tests have meaningful asserts and can be run automatically.

The few that don't aren't mean to be run automatically are in smclient.manual.  Those include
- ApiHandle_Manual and JobClient_FlexUpload_Manual.  Both of these are fully automated, but they can sometimes take a long time due to waiting for file uploads to be processed on the SM side.  These are run from AllManual_Suite.java
- Playground.class, which is made for things that aren't really tests but more like experiments and which may or may not have assert testing, and which may or may not pass

Everything else is a UnitTest of IntegrationTest, made to be run and provide meaningful results automatically

There were a broad set of fixes that generally can be summarized as "improve tests, move to the right class, use actual assert testing everywhere possible".  Things that don't fit that mold:
- fix pluralization of Transactions_IntegrationTest to fit the project standard
- fix misnaming of the projecthealth package (projecthealth -> com.greenfiling.smclient.projecthealth)